🧹 (lwm): Remove legacy RN BLE transport and related dead code
