+++ apps/ledger-live-mobile/package.json
@@ -115,7 +115,6 @@
     "@ledgerhq/lumen-ui-rnative": "catalog:",
     "@ledgerhq/native-ui": "workspace:^",
     "@ledgerhq/react-native-hid": "workspace:^",
-    "@ledgerhq/react-native-hw-transport-ble": "workspace:^",
     "@ledgerhq/types-cryptoassets": "workspace:^",
     "@ledgerhq/types-devices": "workspace:^",
     "@ledgerhq/types-live": "workspace:^",
