+++ apps/ledger-live-mobile/src/components/BleDevicePairingFlow/DmkBleDevicesScanning.tsx
@@ -11,7 +11,7 @@ import {
 import { DeviceModelId } from "@ledgerhq/types-devices";
 import { useSelector } from "~/context/hooks";
 import { bleDevicesSelector } from "~/reducers/ble";
-import { FilterByDeviceModelId } from "~/components/BleDevicePairingFlow/LegacyBleDevicesScanning";
+import { FilterByDeviceModelId } from "./FilterByDeviceModelId";
 import { useIsFocused } from "@react-navigation/core";
 
 type DmkBleDevicesScanningProps = Omit<BleDevicesScanningProps, "devices"> & {
