+++ apps/ledger-live-mobile/src/components/BleDevicePairingFlow/FilterByDeviceModelId.tsx
@@ -1,0 +1,3 @@
+import { DeviceModelId } from "@ledgerhq/types-devices";
+
+export type FilterByDeviceModelId = null | DeviceModelId;
