+++ apps/ledger-live-mobile/src/components/BleDevicePairingFlow/index.tsx
@@ -3,10 +3,8 @@ import { useDispatch } from "~/context/hooks";
 import { Device } from "@ledgerhq/live-common/hw/actions/types";
 
 import RequiresBLE from "../RequiresBLE";
-import BleDevicePairing from "./BleDevicePairing";
 import { addKnownDevice } from "~/actions/ble";
 import type { BleDevicesScanningProps } from "./BleDevicesScanning";
-import type { BleDevicePairingProps } from "./BleDevicePairing";
 import { track } from "~/analytics";
 import { NavigationHeaderBackButton } from "../NavigationHeaderBackButton";
 import { NavigationHeaderCloseButton } from "../NavigationHeaderCloseButton";
@@ -16,9 +14,8 @@ import { DmkBleDevicesScanning } from "./DmkBleDevicesScanning";
 import { DmkBleDevicePairing } from "./DmkBleDevicePairing";
 import { urls } from "~/utils/urls";
 import { Linking } from "react-native";
-import { LegacyBleDevicesScanning } from "~/components/BleDevicePairingFlow/LegacyBleDevicesScanning";
 import { useLocalizedUrl } from "LLM/hooks/useLocalizedUrls";
-import { BleScanningState, useDeviceManagementKitEnabled } from "@ledgerhq/live-dmk-mobile";
+import { BleScanningState } from "@ledgerhq/live-dmk-mobile";
 
 const TIMEOUT_AFTER_PAIRED_MS = 2000;
 
@@ -63,7 +60,7 @@ export type BleDevicePairingFlowProps = {
   /**
    * During pairing step: callback called when the pairing is done and successful
    */
-  onPairingSuccess: BleDevicePairingProps["onPaired"];
+  onPairingSuccess: (device: Device) => void;
 
   /**
    * During pairing step: if true, this component adds the paired device to the list of known devices by LLM
@@ -198,7 +195,6 @@ const BleDevicePairingFlow: React.FC<BleDevicePairingFlowProps> = ({
       setPairingFlowStep("scanning");
     }
   }, [isPaired]);
-  const isDmkTransportEnabled = useDeviceManagementKitEnabled();
 
   const onOpenHelp = useCallback(() => {
     Linking.openURL(pairingUrl);
@@ -262,38 +258,20 @@ const BleDevicePairingFlow: React.FC<BleDevicePairingFlowProps> = ({
   return (
     <RequiresBLE>
       {pairingFlowStep === "pairing" && deviceToPair !== null ? (
-        isDmkTransportEnabled ? (
-          <DmkBleDevicePairing
-            device={deviceToPair}
-            onPaired={onPaired}
-            onRetry={onRetryPairingFlow}
-            onOpenHelp={onOpenHelp}
-          />
-        ) : (
-          <BleDevicePairing
-            deviceToPair={deviceToPair}
-            onPaired={onPaired}
-            onRetry={onRetryPairingFlow}
-            onOpenHelp={onOpenHelp}
-          />
-        )
+        <DmkBleDevicePairing
+          device={deviceToPair}
+          onPaired={onPaired}
+          onRetry={onRetryPairingFlow}
+          onOpenHelp={onOpenHelp}
+        />
       ) : pairingFlowStep === "scanning" ? (
-        isDmkTransportEnabled ? (
-          <DmkBleDevicesScanning
-            filterByDeviceModelId={filterByDeviceModelId}
-            areKnownDevicesDisplayed={areKnownDevicesDisplayed}
-            areKnownDevicesPairable={areKnownDevicesPairable}
-            onDeviceSelect={onDeviceSelect}
-            bleScanningState={bleScanningState}
-          />
-        ) : (
-          <LegacyBleDevicesScanning
-            filterByDeviceModelId={filterByDeviceModelId}
-            areKnownDevicesDisplayed={areKnownDevicesDisplayed}
-            areKnownDevicesPairable={areKnownDevicesPairable}
-            onDeviceSelect={onDeviceSelect}
-          />
-        )
+        <DmkBleDevicesScanning
+          filterByDeviceModelId={filterByDeviceModelId}
+          areKnownDevicesDisplayed={areKnownDevicesDisplayed}
+          areKnownDevicesPairable={areKnownDevicesPairable}
+          onDeviceSelect={onDeviceSelect}
+          bleScanningState={bleScanningState}
+        />
       ) : null}
     </RequiresBLE>
   );
