+++ apps/ledger-live-mobile/src/const/navigation.ts
@@ -27,8 +27,6 @@ export enum ScreenName {
   CurrencySettings = "CurrencySettings",
   EditCurrencyUnits = "EditCurrencyUnits",
   CustomCALRefInput = "CustomCALRefInput",
-  DebugBLE = "DebugBLE",
-  DebugBLEBenchmark = "DebugBLEBenchmark",
   DebugBLEDevicePairing = "DebugBLEDevicePairing",
   DebugConfiguration = "DebugConfiguration",
   DebugCommandSender = "DebugCommandSender",
@@ -102,7 +100,6 @@ export enum ScreenName {
   MyLedgerDevice = "MyLedgerDevice",
   NotificationsSettings = "NotificationsSettings",
   OperationDetails = "OperationDetails",
-  PairDevices = "PairDevices",
   PasswordAdd = "PasswordAdd",
   PasswordRemove = "PasswordRemove",
   Portfolio = "Wallet",
