+++ apps/ledger-live-mobile/src/react-native-hw-transport-ble/index.ts
@@ -1,60 +1,14 @@
 import Config from "react-native-config";
-import BleTransport from "@ledgerhq/react-native-hw-transport-ble";
 import { DeviceManagementKitBLETransport } from "@ledgerhq/live-dmk-mobile";
 import makeMock from "./makeMock";
 import createAPDUMock from "../logic/createAPDUMock";
-import { Observer } from "rxjs";
-import { SchedulerLike } from "rxjs";
-import { TraceContext } from "@ledgerhq/logs";
-import { Device } from "react-native-ble-plx";
-import type {
-  DescriptorEvent,
-  Observer as TransportObserver,
-  Subscription as TransportSubscription,
-} from "@ledgerhq/hw-transport";
-import { HwTransportError } from "@ledgerhq/errors";
-import { DeviceId, DiscoveredDevice } from "@ledgerhq/device-management-kit";
-import { TransportBleDevice } from "@ledgerhq/live-common/ble/types";
-
-interface CommonTransportConstructor {
-  listen: (
-    observer: TransportObserver<DescriptorEvent<TransportBleDevice | null>, HwTransportError>,
-    context?: TraceContext,
-  ) => TransportSubscription;
-  observeState: (
-    observer: Observer<{
-      type: string;
-      available: boolean;
-    }>,
-  ) => TransportSubscription;
-  disconnectDevice: (deviceId: DeviceId, context?: TraceContext) => Promise<void>;
-  setLogLevel: (level: string) => void;
-  open: (
-    deviceOrId: (Device & DiscoveredDevice) | string,
-    timeoutMs?: number,
-    context?: TraceContext,
-    options?: { rxjsScheduler?: SchedulerLike; matchDeviceByName?: string },
-  ) => Promise<BleTransport | DeviceManagementKitBLETransport>;
-}
 
 const names: { [key: string]: string } = {};
 
 /**
- * Retrieves the appropriate BLE transport instance based on environment configuration and feature flags.
- *
- * - If `Config.MOCK` is `true`, it returns a mock transport for testing.
- * - If `isLDMKEnabled` is `true`, it returns `DeviceManagementKitTransport`.
- * - Otherwise, it defaults to `BleTransport`.
- *
- * @param {Object} options - Configuration options.
- * @param {boolean} options.isLDMKEnabled - Flag to enable Device Management Kit transport.
- * @returns {typeof BleTransport} The selected transport instance.
+ * Retrieves the appropriate BLE transport instance based on environment configuration
  */
-const getBLETransport = ({
-  isLDMKEnabled,
-}: {
-  isLDMKEnabled: boolean;
-}): CommonTransportConstructor => {
+const getBLETransport = () => {
   if (Config.MOCK || Config.DETOX) {
     return makeMock({
       // TODO E2E: This could be dynamically set in bridge/server.js
@@ -87,15 +41,9 @@ const getBLETransport = ({
           serviceUUIDs,
         };
       },
-    }) as unknown as CommonTransportConstructor;
+    });
   } else {
-    // when not in MOCK mode, return DeviceManagementKitTransport if enabled,
-    // otherwise BleTransport
-    return (
-      isLDMKEnabled ? DeviceManagementKitBLETransport : BleTransport
-    ) as CommonTransportConstructor;
-    // NOTE: Some method signatures are not compatible between BleTransport and DeviceManagementKitTransport
-    // so we need to cast to CommonTransportConstructor
+    return DeviceManagementKitBLETransport;
   }
 };
